requirements.txt
asgiref==3.11.1
boto3==1.34.144
botocore==1.34.144
Django==5.2.13
django-cors-headers==4.9.0
django-environ==0.13.0
django-rich==2.2.0
django-rich-logging==0.3.0
django-storages==1.14.6
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
jmespath==1.1.0
loguru==0.7.3
markdown-it-py==4.0.0
mdurl==0.1.2
pillow==12.2.0
psycopg2-binary==2.9.11
Pygments==2.20.0
PyJWT==2.12.1
python-dateutil==2.9.0.post0
rich==15.0.0
ruff==0.15.10
s3transfer==0.10.4
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
urllib3==2.6.3
